# src/config/swagger.json
{
  "components": {
...
    "schemas": {
...
      "User": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "firstName": {
            "type": "string",
            "default": ""
          },
          "lastName": {
            "type": "string",
            "default": ""
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "number",
            "format": "double",
            "default": 1
          },
          "organization": {
            "type": "string"
          },
          "staffID": {
            "type": "string"
          },
          "signupCode": {
            "type": "string"
          },
          "chatRooms": {
            "items": {
              "$ref": "#/components/schemas/ChatRoom"
            },
            "type": "array"
          },
          "favourites": {
            "items": {
              "$ref": "#/components/schemas/Favourite"
            },
            "type": "array"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array"
          },
          "cats": {
            "items": {
              "$ref": "#/components/schemas/Cat"
            },
            "type": "array"
          },
          "status": {
            "type": "number",
            "format": "double",
            "default": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "ChatRoom": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "cat": {
            "$ref": "#/components/schemas/Cat"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array"
          },
          "status": {
            "type": "number",
            "format": "double",
            "default": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "Cat": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "type": "string",
            "default": ""
          },
          "breed": {
            "type": "string",
            "default": ""
          },
          "age": {
            "type": "number",
            "format": "double",
            "default": 0
          },
          "description": {
            "type": "string",
            "default": ""
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "image": {
            "type": "string",
            "default": ""
          },
          "chatRooms": {
            "items": {
              "$ref": "#/components/schemas/ChatRoom"
            },
            "type": "array"
          },
          "favourites": {
            "items": {
              "$ref": "#/components/schemas/Favourite"
            },
            "type": "array"
          },
          "status": {
            "type": "number",
            "format": "double",
            "default": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "Favourite": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "cat": {
            "$ref": "#/components/schemas/Cat"
          },
          "status": {
            "type": "number",
            "format": "double",
            "default": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "Message": {
        "properties": {
          "id": {
            "type": "number",
            "format": "double"
          },
          "chatRoom": {
            "$ref": "#/components/schemas/ChatRoom"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "text": {
            "type": "string",
            "default": ""
          },
          "status": {
            "type": "number",
            "format": "double",
            "default": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "UserLoginFormValues": {
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "info": {
    "title": "pet-shelter-backend",
    "version": "1.0.1",
    "description": "Pet Shelter Backend API",
    "contact": {}
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/user": {
      "get": {
        "operationId": "GetUsers",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [],
        "parameters": []
      },
      "post": {
        "operationId": "CreateUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "GetUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/User"
                    }
                  ],
                  "nullable": true
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/user/register": {
      "post": {
        "operationId": "RegisterUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "operationId": "LoginUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "properties": {
                        "token": {},
                        "role": {
                          "type": "number",
                          "format": "double"
                        },
                        "email": {
                          "type": "string"
                        },
                        "id": {
                          "type": "number",
                          "format": "double"
                        }
                      },
                      "required": [
                        "token",
                        "role",
                        "email",
                        "id"
                      ],
                      "type": "object"
                    },
                    {
                      "properties": {
                        "role": {},
                        "token": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "token",
                        "email",
                        "id"
                      ],
                      "type": "object"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "User"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLoginFormValues"
              }
            }
          }
        }
      }
    },
...
    "/chatroom": {
      "get": {
        "operationId": "GetChatRooms",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ChatRoom"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": [
          "ChatRoom"
        ],
        "security": [],
        "parameters": []
      },
      "post": {
        "operationId": "CreateChatRoom",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatRoom"
                }
              }
            }
          }
        },
        "tags": [
          "ChatRoom"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRoom"
              }
            }
          }
        }
      }
    },
    "/chatroom/{id}": {
      "get": {
        "operationId": "GetChatRoom",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ChatRoom"
                    }
                  ],
                  "nullable": true
                }
              }
            }
          }
        },
        "tags": [
          "ChatRoom"
        ],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "operationId": "UpdateChatRoom",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "status": {
                          "type": "number",
                          "format": "double"
                        },
                        "messages": {
                          "items": {
                            "$ref": "#/components/schemas/Message"
                          },
                          "type": "array"
                        },
                        "cat": {
                          "$ref": "#/components/schemas/Cat"
                        },
                        "user": {
                          "$ref": "#/components/schemas/User"
                        },
                        "id": {
                          "type": "number",
                          "format": "double"
                        }
                      },
                      "type": "object"
                    },
                    {
                      "$ref": "#/components/schemas/ChatRoom"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "ChatRoom"
        ],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRoom"
              }
            }
          }
        }
      }
    },
    "/favourite": {
      "get": {
        "operationId": "GetFavourites",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Favourite"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "tags": [
          "Favourite"
        ],
        "security": [],
        "parameters": []
      },
      "post": {
        "operationId": "CreateFavourite",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Favourite"
                }
              }
            }
          }
        },
        "tags": [
          "Favourite"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Favourite"
              }
            }
          }
        }
      }
    },
    "/favourite/{id}": {
      "get": {
        "operationId": "GetFavourite",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Favourite"
                    }
                  ],
                  "nullable": true
                }
              }
            }
          }
        },
        "tags": [
          "Favourite"
        ],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "operationId": "UpdateFavourite",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "status": {
                          "type": "number",
                          "format": "double"
                        },
                        "cat": {
                          "$ref": "#/components/schemas/Cat"
                        },
                        "user": {
                          "$ref": "#/components/schemas/User"
                        },
                        "id": {
                          "type": "number",
                          "format": "double"
                        }
                      },
                      "type": "object"
                    },
                    {
                      "$ref": "#/components/schemas/Favourite"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Favourite"
        ],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Favourite"
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}